+++ pychiver/archiver.py
@@ -31,8 +31,10 @@ import warnings
 
 import numpy
 
-from .calculations import LinearInterpolationStrategy, alignDataFrames, calculateMovingAverage, findCloseTimestamps, Edge
+from .calculations import LinearInterpolationStrategy, alignDataFrames, calculateMovingAverage, findCloseTimestamps, \
+    Edge
 from .endpoints import JsonEndPointArchiver
+from .domain import PVMetaInfo
 import pandas
 
 
@@ -55,7 +57,7 @@ class Archiver:
 
         self.archiver = DefaultEndPoint(archiver_url=archiver_url)
 
-    def get(self, PV, start_date, end_date=None, entries_limit=None, verbose=False):
+    def get(self, PV, start_date, end_date=None, entries_limit=None, verbose=False, force_non_archived=False):
         """
         Returns the archiver data for one or many pvs within the given start_date and end_date.
 
@@ -64,6 +66,8 @@ class Archiver:
         :param end_date: default None => now()
         :param entries_limit: default None, ie. all entries are extracted
         :param verbose: default False, if True the processing printout is provided
+        :param force_non_archived: default False, when True, an attempt to extract archived data is made,
+                may raise exception
         :return: dict of PV to a DataFrame
         """
         useSeparateLimits = False
@@ -78,15 +82,18 @@ class Archiver:
                 if useSeparateLimits:
                     e_limit = entries_limit[index]
                 dataToReturn[onePV] = self._get(onePV, start_date=start_date, end_date=end_date,
-                                                entries_limit=e_limit, verbose=verbose)[0]
+                                                entries_limit=e_limit, verbose=verbose,
+                                                force_non_archived=force_non_archived)[0]
             return dataToReturn
         else:
             if isinstance(entries_limit, tuple) and len(entries_limit) > 0:
                 entries_limit = entries_limit[0]
             return {PV: self._get(PV, start_date=start_date, end_date=end_date,
-                                  entries_limit=entries_limit, verbose=verbose)[0]}
+                                  entries_limit=entries_limit, verbose=verbose,
+                                  force_non_archived=force_non_archived)[0]}
 
-    def getWaveform(self, onePV: str, start_date, end_date=None, verbose=False) -> pandas.DataFrame:
+    def getWaveform(self, onePV: str, start_date, end_date=None, verbose=False,
+                    force_non_archived=False) -> pandas.DataFrame:
         """
         Returns an pandas DataFrame that contains a waveform
 
@@ -94,16 +101,19 @@ class Archiver:
         :param start_date:
         :param end_date: default None => now()
         :param verbose: default False, if True the processing printout is provided
+        :param force_non_archived: default False, when True, an attempt to extract archived data is made anyway
         :return:
         """
         if not isinstance(onePV, str):
             raise ValueError('Only one waveform at the time! Too many or none PVs provided.')
-        return self._get(onePV, start_date=start_date, end_date=end_date, verbose=verbose, waveform_alert=False)[0]
+        return self._get(onePV, start_date=start_date, end_date=end_date, verbose=verbose,
+                         waveform_alert=False, force_non_archived=force_non_archived)[0]
 
     def getAligned(self, PVS: list, start_date, end_date=None,
                    time_base=None, strategy=LinearInterpolationStrategy,
                    entries_limit=None, time_column="secs_nanos", value_columns=("val",),
-                   verbose=False) -> pandas.DataFrame:
+                   verbose=False,
+                   force_non_archived=False) -> pandas.DataFrame:
         """
         Extracts PVs and aligns them to the timestamps of the first PV in the list or separately provided time base.
         Uses the provided InterpolationStrategy (default one LinearInterpolationStrategy)
@@ -119,10 +129,13 @@ class Archiver:
         :param value_columns: optional, default 'val' column will be used,
                      the alignment can be performed for many columns at the same time, provide a tuple.
         :param verbose: default False
+        :param force_non_archived: default False, when True, an attempt to extract archived data is made anyway
+
         :return: a DataFrame with all PVS and their values
         """
         dict_of_dataframes = self.get(PVS, start_date, end_date=end_date,
-                                      entries_limit=entries_limit, verbose=verbose)
+                                      entries_limit=entries_limit, verbose=verbose,
+                                      force_non_archived=force_non_archived)
         if not isinstance(dict_of_dataframes, dict):
             raise ValueError('Wrong data format provided. Dict of pandas.DataFrames expected, {} provided'. \
                              format(dict_of_dataframes.__class__))
@@ -131,7 +144,8 @@ class Archiver:
                                InterpolationStrategyImpl=strategy, verbose=verbose)
 
     def getMovingAverage(self, PV, start_date, end_date=None, entries_limit=None, window=10,
-                         time_column="secs_nanos", value_columns=("val",), verbose=False) -> pandas.DataFrame:
+                         time_column="secs_nanos", value_columns=("val",),
+                         force_non_archived=False, verbose=False) -> pandas.DataFrame:
         """
         Retrieves the data for a given PV and calculates the moving average for a selected window.
         The resulting dataframe is cleared from all NaN cases.
@@ -143,13 +157,15 @@ class Archiver:
         :param window:
         :param time_column:
         :param value_columns:
+        :param force_non_archived:
         :param verbose:
         :return:
         """
         if isinstance(PV, list) or isinstance(PV, tuple) or isinstance(PV, dict):
             raise ValueError('Cannot handle more than one PV at the time. \
                                 Use getAligned together with calculations.calculateMovingAverage')
-        df, isWaveform = self._get(PV, start_date, end_date=end_date, entries_limit=entries_limit, verbose=verbose)
+        df, isWaveform = self._get(PV, start_date, end_date=end_date, entries_limit=entries_limit,
+                                   force_non_archived=force_non_archived, verbose=verbose)
         if isWaveform:
             warnings.warn('Moving average over the waveform is not implemented! Returning simple DataForm')
             return df
@@ -158,9 +174,10 @@ class Archiver:
                                       verbose=verbose)
 
     def compare(self, PVs: list, start_date, end_date=None,
-                #index_of_reference_PV=0,
+                # index_of_reference_PV=0,
                 compare_edge=Edge.FALLING,
                 tolerance_in_seconds=0.1,
+                force_non_archived=False,
                 verbose=False) -> numpy.array:
         """
         Returns timestamps of the close occurrences of the RISING or FALLING for two
@@ -170,6 +187,7 @@ class Archiver:
         :param end_date:
         :param compare_edge: compare the occurrences of the selected change
         :param tolerance_in_seconds: absolute (+/-) acceptable difference for simultaneous events
+        :param force_non_archived:
         :param verbose:
         :return:
         """
@@ -177,7 +195,8 @@ class Archiver:
             raise ValueError('Only two signals expected for comparison!')
 
         # TODO see what to pass more, limits? etc..
-        dfs = self.get(PVs, start_date=start_date, end_date=end_date, verbose=verbose)
+        dfs = self.get(PVs, start_date=start_date, end_date=end_date, verbose=verbose,
+                       force_non_archived=force_non_archived,)
 
         closeTimeStamps = findCloseTimestamps(dfs,
                                               tolerance_in_seconds=tolerance_in_seconds,
@@ -195,20 +214,24 @@ class Archiver:
         # TODO waveforms should go as ArchiverCollectors
         raise NotImplementedError("Not implemented yet")
 
-    def check(self, PV: str) -> dict:
+    def check(self, PV: str, type=PVMetaInfo.STATUS) -> dict:
         """
         Provides information on the requested PV(s)
-        :param PV:
+        :param type: STATUS (default) returns info on PVs, INFO returns info on a given PVs
+        :param PV: PVs to check status or info
         :return: dict of PV to its data
         """
-        return self.archiver.getPVStatus(PV)
+        return self.archiver.getPVStatus(PV, type=type)
 
-    def _get(self, onePV: str, start_date, end_date=None, entries_limit: int = None, verbose=False, waveform_alert=True) \
+    def _get(self, onePV: str, start_date, end_date=None, entries_limit: int = None, verbose=False,
+             waveform_alert=True, force_non_archived=False) \
             -> pandas.DataFrame:
         isWaveform = False
         status = self.check(onePV)
         df = self.archiver.getEmptyResult()
-        if 'Not' in status[onePV]['status']:
+        if force_non_archived:
+            warnings.warn('Skipping the check if {} is being archived in {}'.format(onePV, self.archiver_url))
+        if 'Not' in status[onePV]['status'] and not force_non_archived:
             warnings.warn('Requested PV \'{}\' is NOT archived in {}, empty dataset will be returned.'
                           .format(onePV, self.archiver_url))
             pass
@@ -219,8 +242,9 @@ class Archiver:
         try:
             if len(df) > 0 and len(df['val'][0]):
                 isWaveform = True
-                if waveform_alert: warnings.warn("The PV \'{}\' you have extracted is type of WAVEFORM with {} samples".
-                                                 format(onePV, len(df['val'][0])))
+                if waveform_alert:
+                    warnings.warn("The PV \'{}\' you have extracted is type of WAVEFORM with {} samples".
+                                  format(onePV, len(df['val'][0])))
         except TypeError:
             pass  # This error is thrown on scalar types, due to len(df['val'])